archive/mirav1/requirements.txt
appnope==0.1.2
attrs==22.1.0
azure-core==1.9.0
azure-storage-blob==12.6.0
backcall==0.2.0
boto3==1.26.4
botocore==1.29.4
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
contextlib2==21.6.0
cryptography==3.3.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.6
dnspython==2.2.1
docutils==0.15.2
google-pasta==0.2.0
idna==2.9
importlib-metadata==4.13.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.17.2
jmespath==0.9.5
jupyter-client==6.1.7
jupyter-core==4.7.0
kiwisolver==1.1.0
matplotlib==3.2.1
msrest==0.6.19
multiprocess==0.70.14
numpy==1.23.4
oauthlib==3.1.0
packaging==21.3
pandas==1.5.1
parso==0.7.1
pathos==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit==3.0.8
protobuf==3.20.3
protobuf3-to-dict==0.1.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.4
pymongo==4.3.2
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2022.6
pyzmq==20.0.0
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.6.0
sagemaker==2.116.0
schema==0.7.5
six==1.14.0
smdebug-rulesconfig==1.0.1
tornado==6.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
zipp==3.10.0
